feat: update chat bot js files structure
