+++ public/js/modules/chat_bot_fns.js
@@ -1,9 +1,21 @@
 "use_strict"
 
 const chatWindow = document.getElementById("chat-window");
+const chatMessages = document.getElementById('chat-messages');
 
 function toggleChat() {
     chatWindow.classList.toggle('hidden');    
 }
 
-export { toggleChat };
+function appendMessage(text, side) {
+    const div = document.createElement('div');
+    div.className = `message ${side}`;
+    div.textContent = text;
+    chatMessages.appendChild(div);
+    chatMessages.scrollTo({
+        top: chatMessages.scrollHeight,
+        behavior: 'smooth'
+    });
+}
+
+export { toggleChat, appendMessage };
